4.14-集成mybatis官方代码生成器，访问数据库成功

+++ member/src/main/java/com/jeffrey/train/member/controller/MemberController.java
@@ -19,7 +19,7 @@ public class MemberController {
 
 
     @GetMapping("/count")
-    public int count()
+    public Long count()
     {
         return memberService.count();
     }
